requirements.txt
Flask>=3.0,<4.0
Flask-Login>=0.6,<0.7
Flask-Migrate>=4.0,<5.0
Flask-SQLAlchemy>=3.1,<4.0
Flask-WTF>=1.2,<1.3
SQLAlchemy>=2.0,<3.0
PyMySQL>=1.1,<2.0
python-dotenv>=1.0,<2.0
passlib[bcrypt]>=1.7,<2.0
email-validator>=2.0,<3.0
pytest>=8.0,<9.0
pytest-flask>=1.3,<2.0
